backend/requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.5.2
Flask-CORS==4.0.0
pymongo==4.5.0
python-dotenv==1.0.0
Werkzeug==2.3.7
